# internal/swagger/swagger.yaml
basePath: /api
definitions:
  github_com_neu-software-practice_software-practice-backend_internal_dto.ChargeItemRef:
    properties:
      id:
        type: integer
      item_type:
        description: check/inspection/disposal/prescription
        type: string
    required:
    - id
    - item_type
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_dto.ChargeRequest:
    properties:
      case_number:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.ChargeItemRef'
        minItems: 1
        type: array
    required:
    - case_number
    - items
    type: object
...
  github_com_neu-software-practice_software-practice-backend_internal_dto.RefundRequest:
    properties:
      case_number:
        type: string
      items:
        items:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.ChargeItemRef'
        minItems: 1
        type: array
    required:
    - case_number
    - items
    type: object
...
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body:
    properties:
      data: {}
      error:
        $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.ErrorBody'
      meta:
        $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Meta'
      success:
        type: boolean
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.ErrorBody:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
    type: object
info:
  contact: {}
  description: 门诊全流程后端：挂号收费 / 门诊医生 / 检查 / 检验 / 处置 / 药房。
  title: 东软云医院 HIS 门诊管理 API
  version: "1.0"
paths:
...
  /charge-records:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        type: string
      - description: 挂号ID
        in: query
        name: register_id
        type: integer
      - description: 动作(收费/退费)
        in: query
        name: action
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 费用记录查询 (F1-5 财务 / F2-11 门诊)
      tags:
      - charge
  /charges:
    post:
      consumes:
      - application/json
      parameters:
      - description: 结算项目
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.ChargeRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 收费结算 (F1-3)
      tags:
      - charge
  /charges/pending:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 待缴费项目 (F1-3)
      tags:
      - charge
  /charges/refund:
    post:
      consumes:
      - application/json
      parameters:
      - description: 退费项目
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_dto.RefundRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 退费 (F1-4)
      tags:
      - charge
  /charges/refund-pending:
    get:
      parameters:
      - description: 病历号
        in: query
        name: case_number
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 可退费项目 (F1-4)
      tags:
      - charge
...
  /departments:
    get:
      parameters:
      - description: 科室类型(财务/门诊/检查/检验/药房/处置)
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 科室列表
      tags:
      - catalog
  /diseases:
    get:
      parameters:
      - description: 编码/名称/ICD 关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 疾病检索 (F2-2)
      tags:
      - catalog
...
  /doctors:
    get:
      parameters:
      - description: 挂号科室ID
        in: query
        name: dept_id
        type: integer
      - description: 挂号级别ID
        in: query
        name: regist_level_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 出诊医生列表 (F1-1)
      tags:
      - catalog
  /drugs:
    get:
      parameters:
      - description: 编码/名称/拼音码关键字
        in: query
        name: keyword
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 药品检索 (F2-9/F5-1)
      tags:
      - catalog
...
  /medical-technologies:
    get:
      parameters:
      - description: 编码/名称关键字
        in: query
        name: keyword
        type: string
      - description: 类型(检查/检验/处置)
        in: query
        name: type
        type: string
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页条数
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 医技项目检索 (F2-3/F2-4/F2-10)
      tags:
      - catalog
...
  /regist-levels:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 挂号级别列表
      tags:
      - catalog
...
  /settle-categories:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_neu-software-practice_software-practice-backend_internal_pkg_response.Body'
      security:
      - BearerAuth: []
      summary: 结算类别列表
      tags:
      - catalog
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"